Project files renamed and reorganized directory structure
