+++ src/SplitPayments.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.29209.152
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "splitpayments", "splitpayments\splitpayments.csproj", "{DE74127D-C53F-4BC1-9AC8-26C639020290}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SplitPayments", "SplitPayments\SplitPayments.csproj", "{DE74127D-C53F-4BC1-9AC8-26C639020290}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
