+++ src/SplitPayments/Controllers/ValuesController.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 
-namespace splitpayments.Controllers
+namespace SplitPayments.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
