+++ src/SplitPayments/Models/Payment.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace splitpayments
+namespace SplitPayments.Models
 {
     public class Payment
     {
