+++ src/SplitPayments/Program.cs
@@ -8,7 +8,7 @@ using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
 
-namespace splitpayments
+namespace SplitPayments
 {
     public class Program
     {
