+++ src/SplitPayments/Startup.cs
@@ -11,7 +11,7 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 
-namespace splitpayments
+namespace SplitPayments
 {
     public class Startup
     {
